# server/docs/swagger.yaml
basePath: /api
definitions:
  handlers.CreateServiceRequest:
    properties:
      duration_minutes:
        minimum: 1
        type: integer
      name:
        type: string
      price:
        minimum: 0
        type: number
    required:
    - duration_minutes
    - name
    - price
    type: object
  handlers.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.UserRole'
        enum:
        - admin
        - customer
    required:
    - email
    - name
    - password
    - role
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handlers.MergeAppointmentsRequest:
    properties:
      services:
        items:
          $ref: '#/definitions/models.Service'
        type: array
    type: object
  handlers.ServiceResponse:
    properties:
      duration_minutes:
        type: integer
      id:
        type: integer
      name:
        type: string
      price:
        type: number
    type: object
  handlers.UpdateServiceRequest:
    properties:
      duration_minutes:
        minimum: 1
        type: integer
      name:
        type: string
      price:
        minimum: 0
        type: number
    required:
    - duration_minutes
    - name
    - price
    type: object
  handlers.UpdateUserRequest:
    properties:
      is_active:
        type: boolean
      name:
        type: string
      phone:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
    type: object
  handlers.UserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      phone:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
    type: object
  models.Appointment:
    properties:
      created_at:
        type: string
      date:
        type: string
      id:
        type: integer
      services:
        items:
          $ref: '#/definitions/models.Service'
        type: array
      status:
        $ref: '#/definitions/models.AppointmentStatus'
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.AppointmentStatus:
    enum:
    - PENDING
    - CONFIRMED
    - DONE
    - CANCELED
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusConfirmed
    - StatusDone
    - StatusCanceled
  models.Service:
    properties:
      duration_minutes:
        type: integer
      id:
        type: integer
      name:
        type: string
      price:
        type: number
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      phone:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
      updated_at:
        type: string
    type: object
  models.UserRole:
    enum:
    - admin
    - customer
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleCustomer
host: localhost:8080
info:
  contact: {}
  description: Hair Salon Management API with JWT Authentication
  title: Hair Salon API
  version: "1.0"
paths:
  /admin/appointments/{id}/confirm:
    post:
      consumes:
      - application/json
      description: Atualiza o status de um agendamento operacionalmente
      parameters:
      - description: ID do agendamento
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Atualiza o status de um agendamento
      tags:
      - admin
  /admin/incoming:
    get:
      consumes:
      - application/json
      description: Listagem operacional de agendamentos recebidos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Appointment'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Lista agendamentos recebidos
      tags:
      - admin
  /admin/services:
    post:
      consumes:
      - application/json
      description: Create a new service
      parameters:
      - description: Service data
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateServiceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.ServiceResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create a new service (admin only)
      tags:
      - services
  /admin/services/{id}:
    delete:
      description: Delete a specific service
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete service (admin only)
      tags:
      - services
    put:
      consumes:
      - application/json
      description: Update a specific service
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated service data
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateServiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ServiceResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update service (admin only)
      tags:
      - services
  /admin/users:
    get:
      description: Retrieve all users from the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.UserResponse'
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: List all users (admin only)
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Create a new user with the provided details
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create a new user (admin only)
      tags:
      - admin
  /admin/users/{id}:
    delete:
      description: Delete a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete user (admin only)
      tags:
      - admin
    get:
      description: Retrieve a specific user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get user by ID (admin only)
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Update a specific user's details
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update user (admin only)
      tags:
      - admin
  /appointments:
    get:
      consumes:
      - application/json
      description: Retorna histórico ou agendamentos em um período
      parameters:
      - description: Data inicial
        in: query
        name: start_date
        type: string
      - description: Data final
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Appointment'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - Bearer: []
      summary: Lista agendamentos
      tags:
      - appointments
    post:
      consumes:
      - application/json
      description: Permite agendar um ou mais serviços
      parameters:
      - description: Dados do agendamento
        in: body
        name: appointment
        required: true
        schema:
          $ref: '#/definitions/models.Appointment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Appointment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - Bearer: []
      summary: Cria um novo agendamento
      tags:
      - appointments
  /appointments/{id}:
    put:
      consumes:
      - application/json
      description: Permite alteração até 2 dias antes
      parameters:
      - description: ID do agendamento
        in: path
        name: id
        required: true
        type: integer
      - description: Dados do agendamento
        in: body
        name: appointment
        required: true
        schema:
          $ref: '#/definitions/models.Appointment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Appointment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - Bearer: []
      summary: Atualiza um agendamento
      tags:
      - appointments
  /appointments/{id}/merge:
    post:
      consumes:
      - application/json
      description: Adiciona novos serviços a um agendamento existente
      parameters:
      - description: ID do agendamento existente
        in: path
        name: id
        required: true
        type: integer
      - description: Novos serviços
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.MergeAppointmentsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Appointment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - Bearer: []
      summary: Mescla serviços em um agendamento existente
      tags:
      - appointments
  /services:
    get:
      description: Retrieve all available services
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.ServiceResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all services
      tags:
      - services
  /services/{id}:
    get:
      description: Retrieve a specific service by ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ServiceResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get service by ID
      tags:
      - services
schemes:
- http
swagger: "2.0"